types: changed step_motor_pos type

+++ Laboratorio_3/include/globals.h
@@ -2,6 +2,6 @@
 
 extern uint8_t step_mode;
 
-extern uint8_t step_motor_pos;
+extern int32_t step_motor_pos;
 
 extern const int32_t kb[3][4];
